LPS-104697 Change namings and adapt correspondingly
