+++ modules/apps/headless/headless-delivery/headless-delivery-test/src/testIntegration/java/com/liferay/headless/delivery/resource/v1_0/test/SitePageResourceTest.java
@@ -15,8 +15,40 @@
 package com.liferay.headless.delivery.resource.v1_0.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
+import com.liferay.headless.delivery.client.dto.v1_0.SitePage;
+import com.liferay.layout.page.template.importer.LayoutPageTemplatesImporter;
+import com.liferay.layout.page.template.model.LayoutPageTemplateStructure;
+import com.liferay.layout.page.template.model.LayoutPageTemplateStructureRel;
+import com.liferay.layout.page.template.service.LayoutPageTemplateStructureLocalService;
+import com.liferay.layout.page.template.service.LayoutPageTemplateStructureRelLocalService;
+import com.liferay.layout.util.structure.LayoutStructure;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.model.Layout;
+import com.liferay.portal.kernel.model.LayoutConstants;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.security.auth.PrincipalThreadLocal;
+import com.liferay.portal.kernel.service.LayoutLocalServiceUtil;
+import com.liferay.portal.kernel.service.ServiceContext;
+import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
+import com.liferay.portal.kernel.test.util.RandomTestUtil;
+import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
+import com.liferay.portal.kernel.test.util.TestPropsValues;
+import com.liferay.portal.kernel.util.HashMapBuilder;
+import com.liferay.portal.kernel.util.LocaleUtil;
+import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.test.rule.Inject;
+import com.liferay.segments.constants.SegmentsEntryConstants;
+import com.liferay.segments.model.SegmentsEntry;
+import com.liferay.segments.model.SegmentsExperience;
+import com.liferay.segments.service.SegmentsEntryLocalService;
+import com.liferay.segments.service.SegmentsExperienceLocalService;
 
+import java.io.InputStream;
+
+import org.junit.Assert;
 import org.junit.Ignore;
+import org.junit.Test;
 import org.junit.runner.RunWith;
 
 /**
@@ -25,4 +57,209 @@ import org.junit.runner.RunWith;
 @Ignore
 @RunWith(Arquillian.class)
 public class SitePageResourceTest extends BaseSitePageResourceTestCase {
+
+	@Override
+	@Test
+	public void testGetSiteSitePageExperienceExperienceKeyRenderedPage()
+		throws Exception {
+
+		Layout layout = _addLayout(testGroup.getGroupId(), true);
+
+		String friendlyURL = layout.getFriendlyURL();
+		SegmentsExperience segmentsExperience = _addSegmentsExperience(
+			layout,
+			ServiceContextTestUtil.getServiceContext(testGroup.getGroupId()));
+
+		String siteSitePageExperienceExperienceKeyRenderedPage =
+			sitePageResource.getSiteSitePageExperienceExperienceKeyRenderedPage(
+				testGroup.getGroupId(), friendlyURL.substring(1),
+				segmentsExperience.getSegmentsExperienceKey());
+
+		Assert.assertNotNull(siteSitePageExperienceExperienceKeyRenderedPage);
+	}
+
+	@Override
+	@Test
+	public void testGetSiteSitePageRenderedPage() throws Exception {
+		Layout layout = _addLayout(testGroup.getGroupId(), true);
+
+		String friendlyURL = layout.getFriendlyURL();
+
+		String siteSitePageRenderedPage =
+			sitePageResource.getSiteSitePageRenderedPage(
+				testGroup.getGroupId(), friendlyURL.substring(1));
+
+		Assert.assertNotNull(siteSitePageRenderedPage);
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testGetSiteSitePagesPageWithSortString() {
+	}
+
+	@Override
+	protected String[] getAdditionalAssertFieldNames() {
+		return new String[] {"title"};
+	}
+
+	@Override
+	protected SitePage testGetSiteSitePagesPage_addSitePage(
+			Long siteId, SitePage sitePage)
+		throws Exception {
+
+		Layout layout = _addLayout(siteId, false);
+
+		sitePage.setDateCreated(layout.getCreateDate());
+		sitePage.setDateModified(layout.getModifiedDate());
+		sitePage.setDatePublished(layout.getPublishDate());
+
+		String friendlyURL = layout.getFriendlyURL();
+
+		sitePage.setFriendlyUrlPath(friendlyURL.substring(1));
+
+		sitePage.setSiteId(siteId);
+		sitePage.setTitle(layout.getName(LocaleUtil.getDefault()));
+		sitePage.setUuid(layout.getUuid());
+
+		return sitePage;
+	}
+
+	private Layout _addLayout(long siteId, boolean importPageDefinition)
+		throws Exception {
+
+		Layout layout = LayoutLocalServiceUtil.addLayout(
+			TestPropsValues.getUserId(), siteId, false,
+			LayoutConstants.DEFAULT_PARENT_LAYOUT_ID,
+			RandomTestUtil.randomString(), RandomTestUtil.randomString(),
+			StringPool.BLANK, LayoutConstants.TYPE_CONTENT, false,
+			StringPool.BLANK, ServiceContextTestUtil.getServiceContext(siteId));
+
+		if (importPageDefinition) {
+			String name = PrincipalThreadLocal.getName();
+
+			try {
+				PrincipalThreadLocal.setName(TestPropsValues.getUserId());
+
+				ServiceContext serviceContext =
+					ServiceContextTestUtil.getServiceContext(
+						testGroup.getGroupId());
+
+				ServiceContextThreadLocal.pushServiceContext(serviceContext);
+
+				LayoutPageTemplateStructure layoutPageTemplateStructure =
+					_layoutPageTemplateStructureLocalService.
+						fetchLayoutPageTemplateStructure(
+							testGroup.getGroupId(), layout.getPlid(), true);
+
+				LayoutStructure layoutStructure = LayoutStructure.of(
+					layoutPageTemplateStructure.getData(0));
+
+				layoutStructure.addRootLayoutStructureItem();
+
+				_layoutPageTemplatesImporter.importPageElement(
+					layout, layoutStructure, layoutStructure.getMainItemId(),
+					_read("test-content-page-definition.json"), 0);
+			}
+			finally {
+				PrincipalThreadLocal.setName(name);
+
+				ServiceContextThreadLocal.popServiceContext();
+			}
+		}
+
+		return layout;
+	}
+
+	private SegmentsExperience _addSegmentsExperience(
+			Layout layout, ServiceContext serviceContext)
+		throws Exception {
+
+		SegmentsEntry segmentsEntry =
+			_segmentsEntryLocalService.addSegmentsEntry(
+				null,
+				HashMapBuilder.put(
+					LocaleUtil.getDefault(), RandomTestUtil.randomString()
+				).build(),
+				null, true, null, User.class.getName(), serviceContext);
+
+		SegmentsExperience segmentsExperience =
+			_segmentsExperienceLocalService.addSegmentsExperience(
+				segmentsEntry.getSegmentsEntryId(),
+				_portal.getClassNameId(Layout.class.getName()),
+				layout.getPlid(),
+				HashMapBuilder.put(
+					LocaleUtil.getDefault(), RandomTestUtil.randomString()
+				).build(),
+				true, serviceContext);
+
+		LayoutPageTemplateStructure layoutPageTemplateStructure =
+			_layoutPageTemplateStructureLocalService.
+				fetchLayoutPageTemplateStructure(
+					testGroup.getGroupId(), layout.getPlid());
+
+		LayoutPageTemplateStructureRel layoutPageTemplateStructureRel =
+			_layoutPageTemplateStructureRelLocalService.
+				fetchLayoutPageTemplateStructureRel(
+					layoutPageTemplateStructure.
+						getLayoutPageTemplateStructureId(),
+					SegmentsEntryConstants.ID_DEFAULT);
+
+		layoutPageTemplateStructureRel.setSegmentsExperienceId(
+			segmentsExperience.getSegmentsExperienceId());
+
+		_layoutPageTemplateStructureRelLocalService.
+			updateLayoutPageTemplateStructureRel(
+				layoutPageTemplateStructureRel);
+
+		return segmentsExperience;
+	}
+
+	/*private SitePage _addSitePage() throws Exception {
+		Layout layout = _addLayout(testGroup.getGroupId(), false);
+
+		String friendlyURL = layout.getFriendlyURL();
+
+		return new SitePage() {
+			{
+				dateCreated = layout.getCreateDate();
+				dateModified = layout.getModifiedDate();
+				datePublished = layout.getLastPublishDate();
+				friendlyUrlPath = friendlyURL.substring(1);
+				siteId = layout.getGroupId();
+				title = layout.getName(LocaleUtil.getDefault());
+				uuid = layout.getUuid();
+			}
+		};
+	}*/
+
+	private String _read(String fileName) throws Exception {
+		Class<?> clazz = getClass();
+
+		InputStream inputStream = clazz.getResourceAsStream(
+			"dependencies/" + fileName);
+
+		return StringUtil.read(inputStream);
+	}
+
+	@Inject
+	private LayoutPageTemplatesImporter _layoutPageTemplatesImporter;
+
+	@Inject
+	private LayoutPageTemplateStructureLocalService
+		_layoutPageTemplateStructureLocalService;
+
+	@Inject
+	private LayoutPageTemplateStructureRelLocalService
+		_layoutPageTemplateStructureRelLocalService;
+
+	@Inject
+	private Portal _portal;
+
+	@Inject
+	private SegmentsEntryLocalService _segmentsEntryLocalService;
+
+	@Inject
+	private SegmentsExperienceLocalService _segmentsExperienceLocalService;
+
 }
